requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
discord.py==1.3.4
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.0
multidict==4.7.6
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.5.0
